# Add a JIT warm-up in order to have a fair cache no-cache comparison