before
```
=== Measuring Triton Flash Attention Kernel Startup Time ===
[Cold Start] Triton kernel first run took 3.959432 seconds
[Warm Start] Triton kernel cached run took 0.000113 seconds
Speedup from caching: 35184.46x faster
```

after adding a warm up and deleting triton cache

```
=== Measuring Triton Flash Attention Kernel Startup Time ===
Successfully cleared Triton cache at: /home/fulvius/.triton/cache
[Cold-no-cache Start] Triton kernel without cache  run took 0.000133 seconds
[Warm-cache Start] Triton kernel cached run took 0.000078 seconds
Speedup from caching: 1.70x faster
```